tests: Move RunVMIAndExpectScheduling to libclient
Also move RunVMIAndExpectSchedulingWithWarningPolicy that is called by
it.
