+++ tests/vmi_configuration_test.go
@@ -164,7 +164,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 
 		It("[test_id:3110]lead to get the burstable QOS class assigned when limit and requests differ", func() {
 			vmi := libvmifact.NewAlpine()
-			vmi = tests.RunVMIAndExpectScheduling(vmi, 60)
+			vmi = libclient.RunVMIAndExpectScheduling(vmi, 60)
 
 			Eventually(func() kubev1.PodQOSClass {
 				vmi, err := virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, metav1.GetOptions{})
@@ -193,7 +193,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 
 			By("adding a sidecar to ensure it gets limits assigned too")
 			vmi.ObjectMeta.Annotations = RenderSidecar(kubevirt_hooks_v1alpha2.Version)
-			vmi = tests.RunVMIAndExpectScheduling(vmi, 60)
+			vmi = libclient.RunVMIAndExpectScheduling(vmi, 60)
 
 			Eventually(func() kubev1.PodQOSClass {
 				vmi, err := virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, metav1.GetOptions{})
@@ -219,7 +219,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 
 			By("adding a sidecar to ensure it gets limits assigned too")
 			vmi.ObjectMeta.Annotations = RenderSidecar(kubevirt_hooks_v1alpha2.Version)
-			vmi = tests.RunVMIAndExpectScheduling(vmi, 60)
+			vmi = libclient.RunVMIAndExpectScheduling(vmi, 60)
 
 			Eventually(func() kubev1.PodQOSClass {
 				vmi, err := virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, metav1.GetOptions{})
@@ -257,7 +257,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			vmi := libvmi.New(vmiOptions...)
 
 			By("Starting a VirtualMachineInstance")
-			vmi = tests.RunVMIAndExpectScheduling(vmi, 60)
+			vmi = libclient.RunVMIAndExpectScheduling(vmi, 60)
 			libwait.WaitForSuccessfulVMIStart(vmi)
 
 			expectedMemoryInKiB := expectedGuestMemory * 1024
@@ -1658,7 +1658,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 				vmi := libvmifact.NewGuestless()
 				// Runtime class related warnings are expected since we created a fake runtime class that isn't supported
 				wp := watcher.WarningsPolicy{FailOnWarnings: true, WarningsIgnoreList: []string{"RuntimeClass"}}
-				vmi = tests.RunVMIAndExpectSchedulingWithWarningPolicy(vmi, 30, wp)
+				vmi = libclient.RunVMIAndExpectSchedulingWithWarningPolicy(vmi, 30, wp)
 
 				By("Checking for presence of runtimeClassName")
 				pod, err := libpod.GetPodByVirtualMachineInstance(vmi, vmi.Namespace)
@@ -1702,14 +1702,14 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 				origVmiWithHeadroom := libvmi.New(libvmi.WithResourceMemory(guestMemoryStr))
 
 				By("Running a vmi without additional headroom")
-				vmiWithoutHeadroom := tests.RunVMIAndExpectScheduling(origVmiWithoutHeadroom, 60)
+				vmiWithoutHeadroom := libclient.RunVMIAndExpectScheduling(origVmiWithoutHeadroom, 60)
 
 				By("Setting a headroom ratio in Kubevirt CR")
 				const ratio = "1.567"
 				setHeadroom(ratio)
 
 				By("Running a vmi with additional headroom")
-				vmiWithHeadroom := tests.RunVMIAndExpectScheduling(origVmiWithHeadroom, 60)
+				vmiWithHeadroom := libclient.RunVMIAndExpectScheduling(origVmiWithHeadroom, 60)
 
 				requestWithoutHeadroom := getComputeMemoryRequest(vmiWithoutHeadroom)
 				requestWithHeadroom := getComputeMemoryRequest(vmiWithHeadroom)
@@ -1951,7 +1951,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 				libvmi.WithResourceMemory(enoughMemForSafeBiosEmulation),
 				WithSchedulerName("my-custom-scheduler"),
 			)
-			runningVMI := tests.RunVMIAndExpectScheduling(vmi, 30)
+			runningVMI := libclient.RunVMIAndExpectScheduling(vmi, 30)
 			launcherPod, err := libpod.GetPodByVirtualMachineInstance(runningVMI, testsuite.GetTestNamespace(vmi))
 			Expect(err).ToNot(HaveOccurred())
 			Expect(launcherPod.Spec.SchedulerName).To(Equal("my-custom-scheduler"))
@@ -1965,7 +1965,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 				libvmi.WithResourceMemory(enoughMemForSafeBiosEmulation),
 				libvmi.WithResourceCPU("500m"),
 			)
-			runningVMI := tests.RunVMIAndExpectScheduling(vmi, 30)
+			runningVMI := libclient.RunVMIAndExpectScheduling(vmi, 30)
 
 			readyPod, err := libpod.GetPodByVirtualMachineInstance(runningVMI, testsuite.GetTestNamespace(vmi))
 			Expect(err).ToNot(HaveOccurred())
@@ -1976,7 +1976,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 
 		It("[test_id:3128]should set CPU request when it is not provided", func() {
 			vmi := libvmifact.NewGuestless()
-			runningVMI := tests.RunVMIAndExpectScheduling(vmi, 30)
+			runningVMI := libclient.RunVMIAndExpectScheduling(vmi, 30)
 
 			readyPod, err := libpod.GetPodByVirtualMachineInstance(runningVMI, testsuite.GetTestNamespace(vmi))
 			Expect(err).ToNot(HaveOccurred())
@@ -1994,7 +1994,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			tests.UpdateKubeVirtConfigValueAndWait(config)
 
 			vmi := libvmifact.NewGuestless()
-			runningVMI := tests.RunVMIAndExpectScheduling(vmi, 30)
+			runningVMI := libclient.RunVMIAndExpectScheduling(vmi, 30)
 
 			readyPod, err := libpod.GetPodByVirtualMachineInstance(runningVMI, testsuite.GetTestNamespace(vmi))
 			Expect(err).ToNot(HaveOccurred())
@@ -2017,7 +2017,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 		It("should not set a CPU limit if the namespace doesn't match the selector", func() {
 			By("Creating a running VMI")
 			vmi := libvmifact.NewGuestless()
-			runningVMI := tests.RunVMIAndExpectScheduling(vmi, 30)
+			runningVMI := libclient.RunVMIAndExpectScheduling(vmi, 30)
 
 			By("Ensuring no CPU limit is set")
 			readyPod, err := libpod.GetPodByVirtualMachineInstance(runningVMI, testsuite.GetTestNamespace(vmi))
@@ -2038,7 +2038,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			Expect(err).NotTo(HaveOccurred())
 
 			By("Starting the VMI")
-			runningVMI := tests.RunVMIAndExpectScheduling(vmi, 30)
+			runningVMI := libclient.RunVMIAndExpectScheduling(vmi, 30)
 
 			By("Ensuring the CPU limit is set to the correct value")
 			readyPod, err := libpod.GetPodByVirtualMachineInstance(runningVMI, testsuite.GetTestNamespace(vmi))
@@ -2056,7 +2056,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			It("should not automatically set memory limits in the virt-launcher pod", func() {
 				vmi := libvmifact.NewCirros()
 				By("Creating a running VMI")
-				runningVMI := tests.RunVMIAndExpectScheduling(vmi, 30)
+				runningVMI := libclient.RunVMIAndExpectScheduling(vmi, 30)
 
 				By("Ensuring no memory and cpu limits are set")
 				readyPod, err := libpod.GetPodByVirtualMachineInstance(runningVMI, testsuite.GetTestNamespace(vmi))
@@ -2113,7 +2113,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 
 			It("should set a memory limit in the virt-launcher pod", func() {
 				By("Starting the VMI")
-				runningVMI := tests.RunVMIAndExpectScheduling(vmi, 30)
+				runningVMI := libclient.RunVMIAndExpectScheduling(vmi, 30)
 
 				By("Ensuring the memory and cpu limits are set to the correct values")
 				readyPod, err := libpod.GetPodByVirtualMachineInstance(runningVMI, testsuite.GetTestNamespace(vmi))
@@ -3327,7 +3327,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			vmi.Spec.TopologySpreadConstraints = tsc
 
 			By("Starting a VirtualMachineInstance")
-			vmi = tests.RunVMIAndExpectScheduling(vmi, 30)
+			vmi = libclient.RunVMIAndExpectScheduling(vmi, 30)
 
 			By("Ensuring that pod has expected topologySpreadConstraints")
 			pod, err := libpod.GetPodByVirtualMachineInstance(vmi, vmi.Namespace)
